requirements.txt
strands-agents[ollama]>=1.0.0
python-dotenv>=1.0.0
